+++ src/Stratis.Bitcoin.Features.ProvenHeaders/Behaviors/ProvenHeadersReservedSlotsBehavior.cs
@@ -11,7 +11,7 @@ using Stratis.Bitcoin.P2P.Protocol.Payloads;
 using Stratis.Bitcoin.Utilities;
 using TracerAttributes;
 
-namespace Stratis.Bitcoin.Features.ProvenHeaders
+namespace Stratis.Bitcoin.Features.ProvenHeaders.Behaviors
 {
     public class ProvenHeadersReservedSlotsBehavior : NetworkPeerBehavior
     {
