+++ src/Stratis.Bitcoin.Features.ProvenHeaders/ProvenHeadersFeature.cs
@@ -3,14 +3,11 @@ using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
-using Stratis.Bitcoin.Base;
-using Stratis.Bitcoin.Base.Deployments;
 using Stratis.Bitcoin.Builder.Feature;
 using Stratis.Bitcoin.Connection;
 using Stratis.Bitcoin.Consensus;
 using Stratis.Bitcoin.Features.BlockStore;
-using Stratis.Bitcoin.Features.ProvenHeaders.Store;
-using Stratis.Bitcoin.Interfaces;
+using Stratis.Bitcoin.Features.ProvenHeaders.Behaviors;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.Utilities;
